fix: Obey font styles
Fixes #3795
