+++ Plugins/GitUIPluginInterfaces/GitUIPluginInterfaces.csproj
@@ -69,6 +69,7 @@
     <Compile Include="BuildServerIntegration\IBuildServerTypeMetadata.cs" />
     <Compile Include="BuildServerIntegration\IBuildServerWatcher.cs" />
     <Compile Include="ExecutionResult.cs" />
+    <Compile Include="FontParser.cs" />
     <Compile Include="GitCmdResult.cs" />
     <Compile Include="GitUIPostActionEventArgs.cs" />
     <Compile Include="IBrowseRepo.cs" />
