+++ Plugins/GitUIPluginInterfaces/ISettingsSource.cs
@@ -183,50 +183,4 @@ namespace GitUIPluginInterfaces
             return GetValue(name, defaultValue, x => x);
         }
     }
-
-    public static class FontParser
-    {
-        private const string InvariantCultureId = "_IC_";
-
-        public static string AsString(this Font value)
-        {
-            return string.Format(CultureInfo.InvariantCulture,
-                "{0};{1};{2}", value.FontFamily.Name, value.Size, InvariantCultureId);
-        }
-
-        public static Font Parse(this string value, Font defaultValue)
-        {
-            if (value == null)
-            {
-                return defaultValue;
-            }
-
-            string[] parts = value.Split(';');
-
-            if (parts.Length < 2)
-            {
-                return defaultValue;
-            }
-
-            try
-            {
-                string fontSize;
-                if (parts.Length == 3 && parts[2] == InvariantCultureId)
-                {
-                    fontSize = parts[1];
-                }
-                else
-                {
-                    fontSize = parts[1].Replace(",", CultureInfo.InvariantCulture.NumberFormat.NumberDecimalSeparator);
-                    fontSize = fontSize.Replace(".", CultureInfo.InvariantCulture.NumberFormat.NumberDecimalSeparator);
-                }
-
-                return new Font(parts[0], float.Parse(fontSize, CultureInfo.InvariantCulture));
-            }
-            catch
-            {
-                return defaultValue;
-            }
-        }
-    }
 }
